backend/refactor: FRFS Flows Unification & Fares
