+++ Backend/app/rider-platform/rider-app/Main/spec/Storage/BecknConfig.yaml
@@ -8,7 +8,7 @@ imports:
   VehicleCategory: BecknV2.OnDemand.Enums
   Network: BecknV2.FRFS.Enums
   EBIXConfig: Domain.Types.Extra.IntegratedBPPConfig
-  CUMTAConfig: Domain.Types.Extra.IntegratedBPPConfig
+  DIRECTConfig: Domain.Types.Extra.IntegratedBPPConfig
   CMRLConfig: Domain.Types.Extra.IntegratedBPPConfig
 
 BecknConfig:
@@ -100,7 +100,7 @@ IntegratedBPPConfig:
 
   types:
     ProviderConfig:
-      enum: "EBIX EBIXConfig, CUMTA CUMTAConfig, CMRL CMRLConfig"
+      enum: "EBIX EBIXConfig, DIRECT DIRECTConfig, CMRL CMRLConfig"
       derive': "Generic, FromJSON, ToJSON, Eq"
 
   fromTType:
