+++ Backend/app/rider-platform/rider-app/Main/src-read-only/Domain/Types/IntegratedBPPConfig.hs
@@ -27,6 +27,6 @@ data IntegratedBPPConfig = IntegratedBPPConfig
 
 data ProviderConfig
   = EBIX Domain.Types.Extra.IntegratedBPPConfig.EBIXConfig
-  | CUMTA Domain.Types.Extra.IntegratedBPPConfig.CUMTAConfig
+  | DIRECT Domain.Types.Extra.IntegratedBPPConfig.DIRECTConfig
   | CMRL Domain.Types.Extra.IntegratedBPPConfig.CMRLConfig
   deriving (Generic, FromJSON, ToJSON, Eq)
