+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/Beckn/FRFS/Common.hs
@@ -57,7 +57,7 @@ data DTicketPayload = DTicketPayload
     toRouteProviderCode :: Text,
     adultQuantity :: Int,
     childQuantity :: Int,
-    busTypeProviderCode :: Text,
+    vehicleTypeProviderCode :: Text,
     expiry :: Text,
     ticketNumber :: Text,
     ticketAmount :: Money,
