+++ Backend/app/rider-platform/rider-app/Main/src/ExternalBPP/CallAPI.hs
@@ -7,7 +7,6 @@ import qualified Beckn.ACL.FRFS.Search as ACL
 import qualified Beckn.ACL.FRFS.Utils as Utils
 import qualified BecknV2.FRFS.Enums as Spec
 import BecknV2.FRFS.Utils
-import BecknV2.OnDemand.Enums
 import Domain.Action.Beckn.FRFS.Common
 import qualified Domain.Action.Beckn.FRFS.OnCancel as DOnCancel
 import qualified Domain.Action.Beckn.FRFS.OnInit as DOnInit
@@ -18,9 +17,9 @@ import Domain.Types.FRFSSearch as DSearch
 import qualified Domain.Types.FRFSTicketBooking as DBooking
 import Domain.Types.Merchant
 import Domain.Types.MerchantOperatingCity
+import Domain.Types.Person
 import Environment
-import qualified ExternalBPP.Bus.Flow as BusFlow
-import qualified ExternalBPP.Metro.Flow as MetroFlow
+import qualified ExternalBPP.Flow as Flow
 import Kernel.External.Types (ServiceFlow)
 import Kernel.Prelude
 import Kernel.Storage.Esqueleto.Config
@@ -29,6 +28,7 @@ import Kernel.Types.Id
 import Kernel.Utils.Common
 import Lib.Payment.Storage.Beam.BeamFlow
 import qualified SharedLogic.CallFRFSBPP as CallFRFSBPP
+import qualified SharedLogic.FRFSUtils as FRFSUtils
 import qualified Storage.CachedQueries.FRFSConfig as CQFRFSConfig
 import Storage.CachedQueries.IntegratedBPPConfig as QIBC
 import qualified Storage.CachedQueries.Station as QStation
@@ -58,8 +58,8 @@ type FRFSConfirmFlow m r =
 search :: FRFSSearchFlow m r => Merchant -> MerchantOperatingCity -> BecknConfig -> DSearch.FRFSSearch -> m ()
 search merchant merchantOperatingCity bapConfig searchReq = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory searchReq.vehicleType) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (_, Nothing) -> do
+  case integratedBPPConfig of
+    Nothing -> do
       fork ("FRFS ONDC SearchReq for " <> show bapConfig.vehicleCategory) $ do
         fromStation <- QStation.findById searchReq.fromStationId >>= fromMaybeM (StationNotFound searchReq.fromStationId.getId)
         toStation <- QStation.findById searchReq.toStationId >>= fromMaybeM (StationNotFound searchReq.toStationId.getId)
@@ -67,15 +67,10 @@ search merchant merchantOperatingCity bapConfig searchReq = do
         logDebug $ "FRFS SearchReq " <> encodeToText bknSearchReq
         Metrics.startMetrics Metrics.SEARCH_FRFS merchant.name searchReq.id.getId merchantOperatingCity.id.getId
         void $ CallFRFSBPP.search bapConfig.gatewayUrl bknSearchReq merchant.id
-    (METRO, Just config) -> do
-      fork "FRFS External METRO SearchReq" $ do
-        onSearchReq <- MetroFlow.search merchant merchantOperatingCity config bapConfig searchReq
-        processOnSearch onSearchReq
-    (BUS, Just _config) -> do
+    Just config -> do
       fork "FRFS External SearchReq" $ do
-        onSearchReq <- BusFlow.search merchant merchantOperatingCity bapConfig searchReq
+        onSearchReq <- Flow.search merchant merchantOperatingCity config bapConfig searchReq
         processOnSearch onSearchReq
-    _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
   where
     processOnSearch :: FRFSSearchFlow m r => DOnSearch.DOnSearch -> m ()
     processOnSearch onSearchReq = do
@@ -85,20 +80,16 @@ search merchant merchantOperatingCity bapConfig searchReq = do
 init :: FRFSConfirmFlow m r => Merchant -> MerchantOperatingCity -> BecknConfig -> (Maybe Text, Maybe Text) -> DBooking.FRFSTicketBooking -> m ()
 init merchant merchantOperatingCity bapConfig (mRiderName, mRiderNumber) booking = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory booking.vehicleType) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (METRO, Nothing) -> do
+  case integratedBPPConfig of
+    Nothing -> do
       providerUrl <- booking.bppSubscriberUrl & parseBaseUrl & fromMaybeM (InvalidRequest "Invalid provider url")
       bknInitReq <- ACL.buildInitReq (mRiderName, mRiderNumber) booking bapConfig Utils.BppData {bppId = booking.bppSubscriberId, bppUri = booking.bppSubscriberUrl} merchantOperatingCity.city
       logDebug $ "FRFS InitReq " <> encodeToText bknInitReq
       Metrics.startMetrics Metrics.INIT_FRFS merchant.name booking.searchId.getId merchantOperatingCity.id.getId
       void $ CallFRFSBPP.init providerUrl bknInitReq merchant.id
-    (METRO, Just config) -> do
-      onInitReq <- MetroFlow.init merchant merchantOperatingCity config bapConfig (mRiderName, mRiderNumber) booking
+    Just config -> do
+      onInitReq <- Flow.init merchant merchantOperatingCity config bapConfig (mRiderName, mRiderNumber) booking
       processOnInit onInitReq
-    (BUS, Just _config) -> do
-      onInitReq <- BusFlow.init merchant merchantOperatingCity bapConfig (mRiderName, mRiderNumber) booking
-      processOnInit onInitReq
-    _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
   where
     processOnInit :: FRFSConfirmFlow m r => DOnInit.DOnInit -> m ()
     processOnInit onInitReq = do
@@ -108,8 +99,8 @@ init merchant merchantOperatingCity bapConfig (mRiderName, mRiderNumber) booking
 cancel :: Merchant -> MerchantOperatingCity -> BecknConfig -> Spec.CancellationType -> DBooking.FRFSTicketBooking -> Flow ()
 cancel merchant merchantOperatingCity bapConfig cancellationType booking = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory booking.vehicleType) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (METRO, Nothing) -> do
+  case integratedBPPConfig of
+    Nothing -> do
       fork "FRFS ONDC Cancel Req" $ do
         frfsConfig <-
           CQFRFSConfig.findByMerchantOperatingCityId merchantOperatingCity.id
@@ -120,50 +111,36 @@ cancel merchant merchantOperatingCity bapConfig cancellationType booking = do
         bknCancelReq <- ACL.buildCancelReq booking bapConfig Utils.BppData {bppId = booking.bppSubscriberId, bppUri = booking.bppSubscriberUrl} frfsConfig.cancellationReasonId cancellationType merchantOperatingCity.city
         logDebug $ "FRFS CancelReq " <> encodeToText bknCancelReq
         void $ CallFRFSBPP.cancel providerUrl bknCancelReq merchant.id
-    (METRO, Just _config) -> return ()
-    (BUS, Just _config) -> return ()
-    _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
+    Just _config -> return ()
 
 confirm :: (DOrder -> Flow ()) -> Merchant -> MerchantOperatingCity -> BecknConfig -> (Maybe Text, Maybe Text) -> DBooking.FRFSTicketBooking -> Flow ()
 confirm onConfirmHandler merchant merchantOperatingCity bapConfig (mRiderName, mRiderNumber) booking = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory booking.vehicleType) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (METRO, Nothing) -> do
+  case integratedBPPConfig of
+    Nothing -> do
       fork "FRFS ONDC Confirm Req" $ do
         providerUrl <- booking.bppSubscriberUrl & parseBaseUrl & fromMaybeM (InvalidRequest "Invalid provider url")
         bknConfirmReq <- ACL.buildConfirmReq (mRiderName, mRiderNumber) booking bapConfig booking.searchId.getId Utils.BppData {bppId = booking.bppSubscriberId, bppUri = booking.bppSubscriberUrl} merchantOperatingCity.city
         logDebug $ "FRFS ConfirmReq " <> encodeToText bknConfirmReq
         Metrics.startMetrics Metrics.CONFIRM_FRFS merchant.name booking.searchId.getId merchantOperatingCity.id.getId
         void $ CallFRFSBPP.confirm providerUrl bknConfirmReq merchant.id
-    (METRO, Just config) -> do
-      fork "FRFS EBIX Confirm Req" $ do
+    Just config -> do
+      fork "FRFS External Confirm Req" $ do
         frfsConfig <-
           CQFRFSConfig.findByMerchantOperatingCityId merchantOperatingCity.id
             >>= fromMaybeM (InternalError $ "FRFS config not found for merchant operating city Id " <> merchantOperatingCity.id.getId)
-        onConfirmReq <- MetroFlow.confirm merchant merchantOperatingCity frfsConfig config bapConfig (mRiderName, mRiderNumber) booking
+        onConfirmReq <- Flow.confirm merchant merchantOperatingCity frfsConfig config bapConfig (mRiderName, mRiderNumber) booking
         onConfirmHandler onConfirmReq
-    (BUS, Just config) -> do
-      fork "FRFS EBIX Confirm Req" $ do
-        frfsConfig <-
-          CQFRFSConfig.findByMerchantOperatingCityId merchantOperatingCity.id
-            >>= fromMaybeM (InternalError $ "FRFS config not found for merchant operating city Id " <> merchantOperatingCity.id.getId)
-        onConfirmReq <- BusFlow.confirm merchant merchantOperatingCity frfsConfig config bapConfig (mRiderName, mRiderNumber) booking
-        onConfirmHandler onConfirmReq
-    _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
 
 status :: Id Merchant -> MerchantOperatingCity -> BecknConfig -> DBooking.FRFSTicketBooking -> Flow ()
 status merchantId merchantOperatingCity bapConfig booking = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory booking.vehicleType) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (METRO, Nothing) -> do
+  case integratedBPPConfig of
+    Nothing -> do
       void $ CallFRFSBPP.callBPPStatus booking bapConfig merchantOperatingCity.city merchantId
-    (METRO, Just config) -> do
-      onStatusReq <- MetroFlow.status merchantId merchantOperatingCity config bapConfig booking
+    Just config -> do
+      onStatusReq <- Flow.status merchantId merchantOperatingCity config bapConfig booking
       processOnStatus onStatusReq
-    (BUS, Just config) -> do
-      onStatusReq <- BusFlow.status merchantId merchantOperatingCity config bapConfig booking
-      processOnStatus onStatusReq
-    _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
   where
     processOnStatus onStatusReq = do
       let bookingOnStatusReq = DOnStatus.Booking onStatusReq
@@ -173,9 +150,9 @@ status merchantId merchantOperatingCity bapConfig booking = do
 verifyTicket :: Id Merchant -> MerchantOperatingCity -> BecknConfig -> Spec.VehicleCategory -> Text -> Flow ()
 verifyTicket merchantId merchantOperatingCity bapConfig vehicleCategory encryptedQrData = do
   integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory vehicleCategory) >>= return . fmap (.providerConfig)
-  case (bapConfig.vehicleCategory, integratedBPPConfig) of
-    (BUS, Just config) -> do
-      onStatusReq <- BusFlow.verifyTicket merchantId merchantOperatingCity config bapConfig encryptedQrData
+  case integratedBPPConfig of
+    Just config -> do
+      onStatusReq <- Flow.verifyTicket merchantId merchantOperatingCity config bapConfig encryptedQrData
       processOnStatus onStatusReq
     _ -> throwError $ InternalError ("Unsupported FRFS Flow vehicleCategory : " <> show bapConfig.vehicleCategory)
   where
@@ -183,3 +160,10 @@ verifyTicket merchantId merchantOperatingCity bapConfig vehicleCategory encrypte
       let verificationOnStatusReq = DOnStatus.TicketVerification onStatusReq
       (merchant', booking') <- DOnStatus.validateRequest verificationOnStatusReq
       DOnStatus.onStatus merchant' booking' verificationOnStatusReq
+
+getFares :: (Metrics.CoreMetrics m, CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r) => Maybe (Id Person) -> Merchant -> MerchantOperatingCity -> BecknConfig -> Maybe Text -> Text -> Text -> Spec.VehicleCategory -> m [FRFSUtils.FRFSFare]
+getFares riderId merchant merchantOperatingCity bapConfig mbRouteCode startStationCode endStationCode vehicleCategory = do
+  integratedBPPConfig <- QIBC.findByDomainAndCityAndVehicleCategory (show Spec.FRFS) merchantOperatingCity.id (frfsVehicleCategoryToBecknVehicleCategory vehicleCategory) >>= return . fmap (.providerConfig)
+  case integratedBPPConfig of
+    Just config -> Flow.getFares riderId merchant merchantOperatingCity config bapConfig mbRouteCode startStationCode endStationCode
+    Nothing -> return []
