+++ Backend/app/rider-platform/rider-app/Main/src/ExternalBPP/ExternalAPI/Bus/EBIX/Order.hs
@@ -1,4 +1,4 @@
-module ExternalBPP.Bus.ExternalAPI.EBIX.Order where
+module ExternalBPP.ExternalAPI.Bus.EBIX.Order where
 
 import API.Types.UI.FRFSTicketService
 import Data.Aeson
@@ -9,8 +9,8 @@ import qualified Data.UUID as UU
 import Domain.Types.FRFSTicketBooking
 import Domain.Types.IntegratedBPPConfig
 import EulerHS.Types as ET
-import ExternalBPP.Bus.ExternalAPI.EBIX.Auth
-import ExternalBPP.Bus.ExternalAPI.Types
+import ExternalBPP.ExternalAPI.Bus.EBIX.Auth
+import ExternalBPP.ExternalAPI.Types
 import Kernel.Beam.Functions as B
 import Kernel.Prelude
 import Kernel.Storage.Esqueleto.Config
