+++ Backend/app/rider-platform/rider-app/Main/src/Lib/JourneyLeg/Interface.hs
@@ -35,14 +35,14 @@ getFare merchantId merchantOperatingCityId leg = \case
     getFareReq :: TaxiLegRequest <- mkTaxiGetFareReq
     JL.getFare getFareReq
   DTrip.Bus -> do
-    getFareReq :: BusLegRequest <- mkBusGetFareReq
-    JL.getFare getFareReq
+    getFareReq :: Maybe BusLegRequest <- mkBusGetFareReq
+    maybe (return Nothing) JL.getFare getFareReq
   DTrip.Metro -> do
-    getFareReq :: MetroLegRequest <- mkMetroGetFareReq
-    JL.getFare getFareReq
+    getFareReq :: Maybe MetroLegRequest <- mkMetroGetFareReq
+    maybe (return Nothing) JL.getFare getFareReq
   DTrip.Subway -> do
-    getFareReq :: SubwayLegRequest <- mkSubwayGetFareReq
-    JL.getFare getFareReq
+    getFareReq :: Maybe SubwayLegRequest <- mkSubwayGetFareReq
+    maybe (return Nothing) JL.getFare getFareReq
   DTrip.Walk -> do
     getFareReq :: WalkLegRequest <- mkWalkGetFareReq
     JL.getFare getFareReq
@@ -62,32 +62,62 @@ getFare merchantId merchantOperatingCityId leg = \case
               merchantOpCity
             }
 
-    mkBusGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg BusLegRequest m) => m BusLegRequest
+    mkBusGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg BusLegRequest m) => m (Maybe BusLegRequest)
     mkBusGetFareReq = do
-      return $
-        BusLegRequestGetFare $
-          BusLegRequestGetFareData
-            { startLocation = leg.startLocation.latLng,
-              endLocation = leg.endLocation.latLng
-            }
+      merchant <- QMerchant.findById merchantId >>= fromMaybeM (MerchantDoesNotExist merchantId.getId)
+      merchantOpCity <- CQMOC.findById merchantOperatingCityId >>= fromMaybeM (MerchantOperatingCityNotFound merchantOperatingCityId.getId)
+      let mbRouteCode = leg.routeDetails >>= (.gtfsId)
+      let mbStartStopCode = leg.fromStopDetails >>= (.gtfsId)
+      let mbEndStopCode = leg.toStopDetails >>= (.gtfsId)
+      case (mbRouteCode, mbStartStopCode, mbEndStopCode) of
+        (Just routeCode, Just startStopCode, Just endStopCode) ->
+          return $
+            Just $
+              BusLegRequestGetFare $
+                BusLegRequestGetFareData
+                  { startLocation = leg.startLocation.latLng,
+                    endLocation = leg.endLocation.latLng,
+                    ..
+                  }
+        _ -> return Nothing
 
-    mkMetroGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg MetroLegRequest m) => m MetroLegRequest
+    mkMetroGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg MetroLegRequest m) => m (Maybe MetroLegRequest)
     mkMetroGetFareReq = do
-      return $
-        MetroLegRequestGetFare $
-          MetroLegRequestGetFareData
-            { startLocation = leg.startLocation.latLng,
-              endLocation = leg.endLocation.latLng
-            }
+      merchant <- QMerchant.findById merchantId >>= fromMaybeM (MerchantDoesNotExist merchantId.getId)
+      merchantOpCity <- CQMOC.findById merchantOperatingCityId >>= fromMaybeM (MerchantOperatingCityNotFound merchantOperatingCityId.getId)
+      let mbRouteCode = leg.routeDetails >>= (.gtfsId)
+      let mbStartStopCode = leg.fromStopDetails >>= (.gtfsId)
+      let mbEndStopCode = leg.toStopDetails >>= (.gtfsId)
+      case (mbRouteCode, mbStartStopCode, mbEndStopCode) of
+        (Just routeCode, Just startStopCode, Just endStopCode) ->
+          return $
+            Just $
+              MetroLegRequestGetFare $
+                MetroLegRequestGetFareData
+                  { startLocation = leg.startLocation.latLng,
+                    endLocation = leg.endLocation.latLng,
+                    ..
+                  }
+        _ -> return Nothing
 
-    mkSubwayGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg SubwayLegRequest m) => m SubwayLegRequest
+    mkSubwayGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg SubwayLegRequest m) => m (Maybe SubwayLegRequest)
     mkSubwayGetFareReq = do
-      return $
-        SubwayLegRequestGetFare $
-          SubwayLegRequestGetFareData
-            { startLocation = leg.startLocation.latLng,
-              endLocation = leg.endLocation.latLng
-            }
+      merchant <- QMerchant.findById merchantId >>= fromMaybeM (MerchantDoesNotExist merchantId.getId)
+      merchantOpCity <- CQMOC.findById merchantOperatingCityId >>= fromMaybeM (MerchantOperatingCityNotFound merchantOperatingCityId.getId)
+      let mbRouteCode = leg.routeDetails >>= (.gtfsId)
+      let mbStartStopCode = leg.fromStopDetails >>= (.gtfsId)
+      let mbEndStopCode = leg.toStopDetails >>= (.gtfsId)
+      case (mbRouteCode, mbStartStopCode, mbEndStopCode) of
+        (Just routeCode, Just startStopCode, Just endStopCode) ->
+          return $
+            Just $
+              SubwayLegRequestGetFare $
+                SubwayLegRequestGetFareData
+                  { startLocation = leg.startLocation.latLng,
+                    endLocation = leg.endLocation.latLng,
+                    ..
+                  }
+        _ -> return Nothing
 
     mkWalkGetFareReq :: (CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, EncFlow m r, JL.JourneyLeg WalkLegRequest m) => m WalkLegRequest
     mkWalkGetFareReq = do
