+++ Backend/app/rider-platform/rider-app/Main/src/Lib/JourneyLeg/Metro.hs
@@ -33,6 +33,5 @@ instance JT.JourneyLeg MetroLegRequest m where
   getInfo (MetroLegRequestGetInfo req) = CFRFS.getInfo req.searchId req.fallbackFare req.distance req.duration
   getInfo _ = throwError (InternalError "Not supported")
 
-  getFare (MetroLegRequestGetFare _) = do
-    return (Just $ JT.GetFareResponse {estimatedMinFare = HighPrecMoney {getHighPrecMoney = 10}, estimatedMaxFare = HighPrecMoney {getHighPrecMoney = 10}})
+  getFare (MetroLegRequestGetFare MetroLegRequestGetFareData {..}) = CFRFS.getFare merchant merchantOpCity Spec.METRO routeCode startStopCode endStopCode
   getFare _ = throwError (InternalError "Not supported")
