+++ Backend/app/rider-platform/rider-app/Main/src/Lib/JourneyLeg/Types/Metro.hs
@@ -69,5 +69,10 @@ data MetroLegRequest
 
 data MetroLegRequestGetFareData = MetroLegRequestGetFareData
   { startLocation :: LatLngV2,
-    endLocation :: LatLngV2
+    endLocation :: LatLngV2,
+    routeCode :: Text,
+    startStopCode :: Text,
+    endStopCode :: Text,
+    merchant :: DMerchant.Merchant,
+    merchantOpCity :: DMOC.MerchantOperatingCity
   }
